LibWeb: Split StringStyleValue out of StyleValue.{h,cpp}
