+++ Userland/Libraries/LibWeb/CSS/StyleValue.h
@@ -622,28 +622,6 @@ private:
     NonnullOwnPtr<CalcSum> m_expression;
 };
 
-class StringStyleValue : public StyleValueWithDefaultOperators<StringStyleValue> {
-public:
-    static ValueComparingNonnullRefPtr<StringStyleValue> create(String const& string)
-    {
-        return adopt_ref(*new StringStyleValue(string));
-    }
-    virtual ~StringStyleValue() override = default;
-
-    ErrorOr<String> to_string() const override { return m_string; }
-
-    bool properties_equal(StringStyleValue const& other) const { return m_string == other.m_string; }
-
-private:
-    explicit StringStyleValue(String const& string)
-        : StyleValueWithDefaultOperators(Type::String)
-        , m_string(string)
-    {
-    }
-
-    String m_string;
-};
-
 class TextDecorationStyleValue final : public StyleValueWithDefaultOperators<TextDecorationStyleValue> {
 public:
     static ValueComparingNonnullRefPtr<TextDecorationStyleValue> create(
